+++ ddbmock/routes/query.py
@@ -5,22 +5,6 @@ from ddbmock.errors import ValidationException
 
 @load_table
 def query(post, table):
-    #FIXME: this line is a temp workaround
-    if u'RangeKeyCondition' not in post:
-        post[u'RangeKeyCondition'] = None
-    if u'AttributesToGet' not in post:
-        post[u'AttributesToGet'] = []
-    if u'Count' not in post:
-        post[u'Count'] = False
-    if u'Limit' not in post:
-        post[u'Limit'] = None
-    if u'ScanIndexForward' not in post:
-        post[u'ScanIndexForward'] = True
-    if u'ExclusiveStartKey' not in post:
-        post[u'ExclusiveStartKey'] = None
-    if u'ConsistentRead' not in post:
-        post[u'ConsistentRead'] = False
-
     if post[u'AttributesToGet'] and post[u'Count']:
         raise ValidationException("Can filter fields when only count is requested")