+++ ddbmock/validators/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-from voluptuous import Schema, Invalid
+from onctuous import Schema, Invalid
 from importlib import import_module
 from ddbmock.errors import ValidationException
 
@@ -11,18 +11,16 @@ def dynamodb_api_validate(action, post):
         :action: name of the route after translation to underscores
         :post: data to validate
         :return: False when no validator found, True on success
-        :raises: any voluptuous exception
+        :raises: any onctuous exception
     """
     try:
         mod = import_module('.{}'.format(action), __name__)
         schema = getattr(mod, 'post')
     except (ImportError, AttributeError):
-        return False # Fixme: should log
+        return post  # Fixme: should log
 
     try:
         validate = Schema(schema, required=True)
-        validate(post)
+        return validate(post)
     except Invalid as e:
         raise ValidationException(str(e.errors))
-
-    return True