+++ ddbmock/validators/batch_write_item.py
@@ -1,9 +1,19 @@
 # -*- coding: utf-8 -*-
 
-from .types import table_name, batch_write_request
+from .types import table_name, item_schema, get_key_schema, Any
 
 post = {
     u"RequestItems": {
-        table_name: batch_write_request,
+        table_name: [Any(
+            {
+                u"PutRequest": {
+                    u"Item": item_schema,
+                }
+            },{
+                u"DeleteRequest": {
+                    u"Key": get_key_schema,
+                }
+            },
+        )],
     },
 }