+++ ddbmock/validators/list_tables.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from voluptuous import optional
-from .types import table_name, table_page
+from .types import table_name, table_page, Required
 
 post = {
-    optional(u'ExclusiveStartTableName'): table_name,
-    optional(u'Limit'): table_page,
+    Required(u'ExclusiveStartTableName', None): table_name,
+    Required(u'Limit', None): table_page,
 }