+++ ddbmock/validators/put_item.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from .types import table_name, optional, item_schema, return_values, expected_schema
+from .types import table_name, Required, item_schema, return_values, expected_schema
 
 post = {
     u'TableName': table_name,
     u'Item': item_schema,
-    optional(u'Expected'): expected_schema, # It is optional but with a def value
-    optional(u'ReturnValues'): return_values, # It is optional but with a def value
+    Required(u'Expected', {}): expected_schema, # It is optional but with a def value
+    Required(u'ReturnValues', u'NONE'): return_values, # It is optional but with a def value
 }