+++ ddbmock/validators/scan.py
@@ -1,14 +1,14 @@
 # -*- coding: utf-8 -*-
 
 from .types import (
-    table_name, optional, item_schema, consistent_read, limit, scan_filter,
-    attributes_to_get_schema, key_field_value, boolean, get_key_schema)
+    table_name, Required, item_schema, consistent_read, limit, scan_filter,
+    attributes_to_get_schema, key_field_value, Boolean, get_key_schema)
 
 post = {
     u'TableName': table_name,
-    optional(u'ScanFilter'): scan_filter,
-    optional(u'Count'): boolean,
-    optional(u'Limit'): limit,
-    optional(u'ExclusiveStartKey'): get_key_schema,
-    optional(u'AttributesToGet'): attributes_to_get_schema, #FIXME: handle default
+    Required(u'ScanFilter', {}): scan_filter,
+    Required(u'Count', False): Boolean,
+    Required(u'Limit', None): limit,
+    Required(u'ExclusiveStartKey', None): get_key_schema,
+    Required(u'AttributesToGet', []): attributes_to_get_schema,
 }