+++ tests/unit/test_router_init.py
@@ -16,12 +16,13 @@ class TestRouterInit(unittest.TestCase):
         m_module = m_import.return_value
         m_route = m_module.create_table
         m_route.return_value = {}
+        m_validated = m_validate.return_value
 
         router(ACTION, POST)
 
         m_import.assert_called_with("ddbmock.routes.{}".format(ROUTE))
         m_validate.assert_called_with(ROUTE, POST)
-        m_route.assert_called_with(POST)
+        m_route.assert_called_with(m_validated)
 
     def test_do_request_route_404(self):
         from ddbmock.router import router
@@ -31,3 +32,24 @@ class TestRouterInit(unittest.TestCase):
                                 'exist',
                                 router,
                                 ACTION_404, POST)
+
+    @mock.patch("ddbmock.router.import_module")
+    @mock.patch("ddbmock.router.dynamodb_api_validate")
+    def test_internal_server_error(self, m_validate, m_import):
+        from ddbmock.router import router
+        from ddbmock.errors import InternalFailure
+
+        m_module = m_import.return_value
+        m_route = m_module.create_table
+        m_route.return_value = {}
+        m_route.side_effect = ValueError
+        m_validated = m_validate.return_value
+
+        self.assertRaisesRegexp(InternalFailure,
+                                'ValueError',
+                                router,
+                                ACTION, POST)
+
+        m_import.assert_called_with("ddbmock.routes.{}".format(ROUTE))
+        m_validate.assert_called_with(ROUTE, POST)
+        m_route.assert_called_with(m_validated)